Factor out common moe functions in collector
